Weave support

+++ remote/run-docker.sh
@@ -46,7 +46,10 @@ then
     ./ezd/bin/pre-build.sh
 fi
 
-docker rmi ${DOCKER_IMAGE}:${DEPLOY_ENV}  || :
+if [[ $EASYDEPLOY_STATE == "stateless" ]]
+then
+    docker rmi ${DOCKER_IMAGE}:${DEPLOY_ENV}  || :
+fi
 docker pull ${DOCKER_IMAGE}:${DEPLOY_ENV}
 docker ps || (sudo service docker restart; sudo service supervisor restart;  exit -1 )
 docker stop ${COMPONENT}-${OFFSET} || :
